requirements.txt
anthropic==0.50.0
schedule==1.2.1
tweepy==4.14.0
praw==7.7.1
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
pytest==8.0.0
